Fix installation of PowerShell extension
The old Snap-In mechanism to register PowerShell extensions has been
deprecated and should not be used anymore.
Instead use the new Module mechanism which was introduced with
PowerShell 2.0 and is much simpler.
It allows us to drop a good chunk of code and streamlines the
installation (e.g. we don't need to call InstallUtil.exe anymore).

We ship the module in our product directory and update PSModulePath in
the installer accordingly.
For module autoloading to work, we need to ship a module manifest file
and and follow certain naming rules which are detailed in:
https://docs.microsoft.com/en-us/powershell/scripting/developer/module/installing-a-powershell-module

After the installation you can list available modules via:

> Get-Module -List-Available

which should now show our PHPManager module. You can load it using:

> Import-Module PHPManager

Fixes: #33